requirements.txt
alembic==1.7.7
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
login==0.0.6
Mako==1.2.0
MarkupSafe==2.1.0
Pillow==9.0.1
psycopg2==2.9.3
pycparser==2.21
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
stripe==2.67.0
urllib3==1.26.8
Werkzeug==2.0.3
WTForms==3.0.1
